Refactored attribute encoding for attestations by introducing a unified `EncodeAttestationAttributes` interface. Consolidated SD-JWT-VC and mdoc implementations under `out.format`. Removed redundant encoding classes and utilities.
